requirements.txt
fastapi
uvicorn
openai
pydantic
streamlit
requests
python-dotenv
PyMuPDF  # For reading PDF files
